# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.SubscriptionIDResponse:
    properties:
      id:
        type: integer
    type: object
  dto.SubscriptionResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.SubscriptionSumPriceResponse:
    properties:
      sum_price:
        type: integer
    type: object
  dto.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for subscriptions service
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: List subscriptions with filters
      parameters:
      - description: service_name
        in: query
        name: service_name
        type: string
      - description: user_id (uuid)
        in: query
        name: user_id
        type: string
      - description: from (MM-YYYY)
        in: query
        name: from
        type: string
      - description: to (MM-YYYY)
        in: query
        name: to
        type: string
      - description: limit (default 50)
        in: query
        name: limit
        type: integer
      - description: offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create subscription
      parameters:
      - description: subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubscriptionIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete subscription by ID
      parameters:
      - description: subscription ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete subscription by ID
      tags:
      - subscriptions
    get:
      description: Get subscription by ID
      parameters:
      - description: subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Update subscription
      parameters:
      - description: subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/sum:
    get:
      description: Sum price subscriptions
      parameters:
      - description: service_name
        in: query
        name: service_name
        type: string
      - description: user_id (uuid)
        in: query
        name: user_id
        type: string
      - description: from (MM-YYYY)
        in: query
        name: from
        type: string
      - description: to (MM-YYYY)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionSumPriceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Sum price subscriptions
      tags:
      - subscriptions
swagger: "2.0"